requirements.txt
boto3
flask
opencv-python
pillow